# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Service for managing teams and tasks",
        "title": "Team Task Managere",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Maksim",
            "url": "https://github.com/hhullen",
            "email": "hhullen@gmail.com"
        },
        "license": {
            "name": "Creative Commons Attribution-NonCommercial 4.0 International Public License",
            "url": "https://creativecommons.org/licenses/by-nc/4.0/deed.en"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/login": {
            "post": {
                "description": "login user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "login user",
                "parameters": [
                    {
                        "description": "login and password",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/datastruct.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/datastruct.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/datastruct.Status"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/datastruct.Status"
                        }
                    }
                }
            }
        },
        "/refresh": {
            "post": {
                "description": "refresh tokens.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "refresh tokens",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/datastruct.RefreshResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/datastruct.Status"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/datastruct.Status"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "register new user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "register new user",
                "parameters": [
                    {
                        "description": "login and password and info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/datastruct.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/datastruct.RegisterResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/datastruct.Status"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/datastruct.Status"
                        }
                    }
                }
            }
        },
        "/teams": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "list teams the user member of .",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "list teams the user member of",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/datastruct.ListUserTeamsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/datastruct.Status"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/datastruct.Status"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "create new team.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "create new team",
                "parameters": [
                    {
                        "description": "team description",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/datastruct.CreateTeamRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/datastruct.CreateTeamResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/datastruct.Status"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/datastruct.Status"
                        }
                    }
                }
            }
        },
        "/teams/{id}/invite": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "invite user to team .",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "invite user to team",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "team id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "team id and login of invited user",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/datastruct.InviteUserToTeamRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/datastruct.InviteUserToTeamResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/datastruct.Status"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/datastruct.Status"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "datastruct.CreateTeamRequest": {
            "type": "object",
            "required": [
                "description",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "responsible for transactions"
                },
                "name": {
                    "type": "string",
                    "example": "core"
                }
            }
        },
        "datastruct.CreateTeamResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "status message"
                }
            }
        },
        "datastruct.InviteUserToTeamRequest": {
            "type": "object",
            "required": [
                "login"
            ],
            "properties": {
                "login": {
                    "type": "string",
                    "example": "VaKadyk359"
                }
            }
        },
        "datastruct.InviteUserToTeamResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "status message"
                }
            }
        },
        "datastruct.ListUserTeamsResponse": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/datastruct.TeamDescription"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "status message"
                }
            }
        },
        "datastruct.LoginRequest": {
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "login": {
                    "type": "string",
                    "example": "VaKadyk359"
                },
                "password": {
                    "type": "string",
                    "example": "Xldf32Q"
                }
            }
        },
        "datastruct.LoginResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "-"
                },
                "status": {
                    "type": "string",
                    "example": "status message"
                }
            }
        },
        "datastruct.RefreshResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "-"
                },
                "status": {
                    "type": "string",
                    "example": "status message"
                }
            }
        },
        "datastruct.RegisterRequest": {
            "type": "object",
            "required": [
                "login",
                "name",
                "password"
            ],
            "properties": {
                "login": {
                    "type": "string",
                    "example": "VaKadyk359"
                },
                "name": {
                    "type": "string",
                    "example": "Vasilisa"
                },
                "password": {
                    "type": "string",
                    "example": "Xldf32Q"
                }
            }
        },
        "datastruct.RegisterResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "status message"
                }
            }
        },
        "datastruct.Status": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "status message"
                }
            }
        },
        "datastruct.TeamDescription": {
            "type": "object",
            "required": [
                "description",
                "name",
                "team_id"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "responsible for transactions"
                },
                "name": {
                    "type": "string",
                    "example": "core"
                },
                "team_id": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type 'Bearer \u003ctoken\u003e' to authenticate",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}